==> Lessons/07_Lab_TimeSeries.Rmd <==
---
title: "7: Lab - Time Series"
author: "Environmental Data Analytics | John Fay and Luana Lima"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M7
2. Explore the components of times series: trend, seasonal, random
3. Perform trend analysis on another dataset

## Set up

```{r, message = FALSE}
#notes
#time series are special bc they are something over time? we can generate projections perhaps to make a more informed decisions
#3 main components of a time series: seasonality, trend, remainder (usually the hardest part to forecast, not explain by trend or seasonality) 
#this component is the cyclical: how is it different from seasonal?- seasonal is over one calendar year, goes along with season, day or hour, cyclical is also repeated but longer durations, ,perhaps only one or two cycles a year, could be even once every 10 years. we need more data to see these 
#stl fxn receives time series object and identifies the three main components, if you sum 3 components you get original data (that's why its a decomposition)
#challenges: dealing with missing observations. we can't just drop na's because it would mess up the spacing (since they are over time and locked to that time, you can't just skip july for ex) instead of getting rid of them we fill them out (model?) using linear interpolation you just connect the dots. not change the series at all
#if you have really big gaps it is better to just drop the data before the gap
#once you fill in gaps you visualize using a line plot 
#the other thing you look at is overall increase or decrease? is there one? how can you be sure? mann kendall, can't use this for seasonal time series, because it will be confused sine it is going up and done. you have to use seasonal mann kendall! it looks  at each season and then you combine the trends from each season to get an overall test 
#spearman correlation, also cannot handle seasonality and has no variation to include seasonality. to get rid of seasonality, you can just subtract from original data (you have to do this in the assignment). or you can group seasonal data, like months over year... i'm a little confused on this one
#two types of trends: stochastic and deterministic 
#deterministic: easier to observe, constant over time, can model with a regression line
#stochastic: harder to identify, changes over time, could be change in pace, could trend up and down 
#test to identify stochastic trend: ADF- Augmented Dickey Fuller 
#important to identify trends that happened in the past to predict future 
#ACF vs PACF
#acf has info about correlation, but we are only talking about one (and time lol) so this is trying to find how yt is related to yt-1 ...etc. help identify order of model. how relevant are previous observations to current observation (helps forcast)
#pacf: when you think about yt and yt-2 there is influence of yt-1, so this model tries to erase that influence, the interaction in between observations. this is important bc there might be high correlation between two observations due to influence of other observations, helps with order of model as well. 
#frequency is when do you start repeating, yearly, monthly, daily. shorter time interval

library(tidyverse)
library(lubridate)
#install.packages("trend")
library(trend)
#install.packages("zoo")
library(zoo)
#install.packages("Kendall")
library(Kendall)
#install.packages("tseries")
library(tseries)

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## Import Datasets

Today we will work with data from the Climate Change Knowledge Portal from the World Bank Group. More specifically historical rainfall and temperature averages for Brazil. You will find two new data files on folder "/Data/Raw/". One with rainfall named "pr_1901_2016_BRA.csv" and another with temperature named "tas_1901_2016_BRA.csv". The data span the period from 1901 to 2016 in monthly steps. You can download the data [here][https://climateknowledgeportal.worldbank.org/download-data]

Research question: Can you see any changes on the rainfall regime or average temperature over time?

```{r}
#Exercise 1: Import both datasets using the read.csv function.
Rainfall_BR <- read.csv("./Data/Raw/pr_1901_2016_BRA.csv", stringsAsFactors = TRUE)

Temp_BR <- read.csv("./Data/Raw/tas_1901_2016_BRA.csv", stringsAsFactors = TRUE)

#Exercise 2: Tidy the rainfall data sets.
#a Rename the column with Rainfall to get rid of the dots.
#b Note that on both data sets that is a column with the month name and average. Convert it to a Month column by separating the string, i.e., month in one column and the word "Average" on another. Hint: use the separate() function in the pipe.
#c Now you should have a column with Month. Use the paste0() function to paste month and year together and create a date object.
#d Select only the columns of interest: Date and rainfall

#separate is changing one column into 3, the limit is a space (column,c("new column", "new column"), "limit")
#paste combines columns together, make sure to use paste0 when there is no separater?

Rainfall_BR_processed <-
  Rainfall_BR %>% 
  rename( Rainfall = Rainfall....MM.) %>% 
  separate(Statistics,c("Null","Month","Null2")," ") %>% 
  mutate( Date = my(paste0(Month,"-",Year))) %>% 
  select(Date,Rainfall)

#Exercise 3: Repeat exercise 2 to the temperature dataset.
Temp_BR_processed <-
  Temp_BR %>% 
  rename( Temperature_C = Temperature....Celsius.) %>% 
  separate(Statistics,c("Null","Month","Null2")," ") %>% 
  mutate( Date = my(paste0(Month,"-",Year))) %>% 
  select(Date,Temperature_C)  


#Exercise 4: Join the temperature and rainfall into one tidy data frame with 3 columns: Date, Rainfall and Temperature.
BRA_complete <- inner_join(Rainfall_BR_processed,Temp_BR_processed)

#inner_join: if there is a month of rainfall where we don't have temp, r just assigns na but won't delete. use this in the assignment!! you can also use left_join 
#we don't like to use column bind here bc we aren't sure if they are matching up times
           
```


## Initial plots

```{r}

#Exercise 5: Check if there is any missing data for both temperature and rainfall series.

summary(BRA_complete$Rainfall)
summary(BRA_complete$Temperature_C)


#Exercise 6: Plot both series over time on separate plots. Add a trend line using geom_smooth(). Can you say something about the trend just by visualizing the data?


temp_plot<-ggplot(BRA_complete, aes(x=Date, y=Temperature_C)) +
  geom_line() +
  geom_smooth(method="lm")
  labs(x="Date", y="Temperature (C)")
print(temp_plot)
  
#the temperature seems to have an upward trend over time

rainfall_plot<- ggplot(BRA_complete, aes(x=Date, y=Rainfall)) +
  geom_line()+
  geom_smooth(method="lm")
print(rainfall_plot)
#rainfall might have an upward trend? but it's hard to see. there is definitely some seasonal variation 


```

## Creating time series objects

```{r}
#Exercise 7: Create a time series objects for each series using ts(). Make sure you specify the starting point and frequency.

#you can also use 
#fmonth<- month(first(BRA_complete$Date))
#fyear<- year(first(BRA_complete$Date))
#and then plug these in start=c(fyear,fmonth)

temp_ts<-ts(BRA_complete$Temperature_C, start=c(1901,1), frequency=12 )

rainfall_ts<-ts(BRA_complete$Rainfall, start=c(1901,1), frequency = 12)

```


## Decomposing a time series dataset

A given time series can be made up of several component series: 

1. A **seasonal** component, which repeats over a fixed known period (e.g., seasons of the year, months, days of the week, hour of the day)
2. A **trend** component, which quantifies the upward or downward progression over time. The trend component of a time series does not have to be monotonic.
3. An **error** or **random** component, which makes up the remainder of the time series after other components have been accounted for. This component reflects the noise in the dataset. 
4. (optional) A **cyclical** component, which repeats over periods greater than the seasonal component. A good example of this is El Niño Southern Oscillation (ENSO) cycles, which occur over a period of 2-8 years.


The `stl` function decomposes the time series object into its component parts. We must specify that the window for seasonal extraction is either "periodic" or a specific number of at least 7. The decomposition proceeds through a loess (locally estimated scatterplot smoothing) function.

```{r}

# Exercise 8: Use the stl() function to decompose both series. 
# Visualize the decomposed series using plot(). What components can you identify?

temp_decomposed<-stl(temp_ts, s.window = "periodic")
plot(temp_decomposed)

rainfall_decomposed<-stl(rainfall_ts, s.window = "periodic")
plot(rainfall_decomposed)

#can identify all components, seasonal, trend, and remainder 

# Exercise 9: Use the times series object generated with the stl() to extract the 
# components of each series and turn them into data frames.

temp.components<- as.data.frame(temp_decomposed$time.series[,1:3])

temp.components<- mutate(temp.components,
        Observed = BRA_complete$Temperature_C,     
        Date = BRA_complete$Date)

rainfall.components<- as.data.frame(rainfall_decomposed$time.series[,1:3])

rainfall.components<- mutate(rainfall.components,
        Observed = BRA_complete$Rainfall,     
        Date = BRA_complete$Date)

# Exercise 10: Visualize how the trend maps onto the data for both series

#ERROR: Error in FUN(X[[i]], ...) : object 'Observed' not found 
temp_plot2<-ggplot(BRA_complete) +
  geom_line(aes(y=Observed, x=Date, size=0.25))+
  geom_line(aes(y=trend,x=Date), color="c13d75ff")
  
  
print(temp_plot2)
  


rainfall_plot2<- ggplot(BRA_complete, aes(x=Date, y=Rainfall)) +
  geom_line()+
  geom_smooth(method="lm")
print(rainfall_plot)


```


## Trend analysis

Monotonic trends are a gradual shift over time that is consistent in direction. Specific tests for monotonic trend analysis are listed below, with assumptions and tips: 

* **linear regression**: no seasonality, fits the assumptions of a parametric test. Function: `lm`
* **Mann-Kendall**: no seasonality, non-parametric, missing data allowed. Function: `Kendall::MannKendall()` or `trend::mk.test()`
* **Seasonal Mann-Kendall**: seasonality, non-parametric `Kendall::SeasonalMannKendall` or `trend::smk.test()`
* **Spearman Rho**: no seasonality, non-parametric, missing data allowed. Function: `stats::cor.test(method="spearman")`


```{r}

# Exercise 11: Apply one of the trend detection test to the original data set. Remember that the data has seasonal component, which test can be applied in this case?









# Exercise 12: Now remove the seasonality and apply another test. Compare the results with what you obtained on Exercise 9.











```

What would we conclude based on these findings? 

> Answer:
